Cleanup v1 changed code.
