+++ src/python/pants/bin/BUILD
@@ -17,9 +17,9 @@ python_library(
     'src/python/pants/build_graph',
     'src/python/pants/core_tasks',
     'src/python/pants/engine/legacy:address_mapper',
-    'src/python/pants/engine/legacy:change_calculator',
     'src/python/pants/engine/legacy:graph',
     'src/python/pants/engine/legacy:parser',
+    'src/python/pants/engine/legacy:source_mapper',
     'src/python/pants/engine:build_files',
     'src/python/pants/engine:fs',
     'src/python/pants/engine:legacy_engine',
@@ -35,6 +35,7 @@ python_library(
     'src/python/pants/option',
     'src/python/pants/reporting',
     'src/python/pants/pantsd:pants_daemon',
+    'src/python/pants/scm:change_calculator',
     'src/python/pants/scm/subsystems:changed',
     'src/python/pants/subsystem',
     'src/python/pants/task',
