+++ src/python/pants/init/BUILD
@@ -15,7 +15,6 @@ python_library(
     'src/python/pants/build_graph',
     'src/python/pants/core_tasks',
     'src/python/pants/engine:native',
-    'src/python/pants/engine/legacy:change_calculator',
     'src/python/pants/goal',
     'src/python/pants/goal:run_tracker',
     'src/python/pants/logging',
